Move Carousel component in UI folder
